todobackend/requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-debug-toolbar==2.1
django-jsonfield==1.4.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
mysqlclient==2.1.1
PyJWT==1.7.1
pyotp==2.7.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
xlrd==1.2.0
xlwt==1.3.0
